A fancy top proposal that matches the tobacco tin
This consists of two parts, the first a topper to glue on that features parts of the tobacco tin design. The topper is 'divided' into 5 levels to be able to be able to perform filament swaps of different colors. The bottom level of the topper is 6mm giving 3, 2mm layer height passes. All other levels are 4mm allowing 2, 2mm layer height passes. The second part is a new top that still fits the box bottom, but has corner bevels that match the topper.