2D countersheet with automation
Added automation script make_countersheet.sh.
Added PNG-version of plane-model, because using
the SVG in the countersheet did not look good
for some reason.

Also fixed the size on counters that was off and
replaced the altitude and health markers with
more useful versions.

With the plane-models as external images it is
no longer possible to set the colors using
the countersheetsextension the way it was
done in the previous versions. That is why
the background octagon was added to the
counters.

Signed-off-by: Pelle Nilsson <pelle@hexandcounter.org>